Поменял тип поля Accrual у структуры Order

+++ internal/models/order.go
@@ -18,6 +18,6 @@ type Orders []Order
 type Order struct {
 	ID         uint64      `json:"number"`
 	Status     OrderStatus `json:"status"`
-	Accrual    uint64      `json:"accrual,omitempty"`
+	Accrual    float64     `json:"accrual,omitempty"`
 	UploadedAt time.Time   `json:"uploaded_at"`
 }
